+++ packages/storybook/src/spright/chat/conversation/chat-conversation-matrix.stories.ts
@@ -10,6 +10,7 @@ import {
 } from '@ni/nimble-components/dist/esm/theme-provider/design-tokens';
 import { buttonTag } from '@ni/nimble-components/dist/esm/button';
 import { iconThumbUpTag } from '@ni/nimble-components/dist/esm/icons/thumb-up';
+import { ChatConversationAppearance } from '@ni/spright-components/dist/esm/chat/conversation/types';
 import { hiddenWrapper } from '../../../utilities/hidden';
 import { createStory } from '../../../utilities/storybook';
 import {
@@ -57,6 +58,12 @@ const contentHeightStates = [
 ] as const;
 type ContentHeightStates = (typeof contentHeightStates)[number];
 
+const appearanceStates = [
+    ['default', ChatConversationAppearance.default],
+    ['overlay', ChatConversationAppearance.overlay]
+] as const;
+type AppearanceStates = (typeof appearanceStates)[number];
+
 const componentSizing = (
     [_messageTypeLabel, messageType]: MessageTypeStates,
     [viewportLabel, viewportWidth, viewportHeight]: ViewportStates,
@@ -70,7 +77,7 @@ const componentSizing = (
         margin-bottom: 0px;
         "
     >
-        viewport:${() => viewportLabel}, content:${() => contentWidthLabel},${() => contentHeightLabel} 
+        viewport:${() => viewportLabel}, content:${() => contentWidthLabel},${() => contentHeightLabel}
     </p>
     <div style="
         width: ${viewportWidth};
@@ -216,6 +223,34 @@ const conversationWithInput = (
 
 export const conversationWithInputSizing: StoryFn = createMatrixThemeStory(html`
     ${createMatrix(conversationWithInput, [heightStates])}
+`);
+
+const conversationWithAppearance = ([
+    appearanceLabel,
+    appearance
+]: AppearanceStates): ViewTemplate => html`
+    <p         
+        style="
+        font: var(${bodyFont.cssCustomProperty});
+        color: var(${bodyFontColor.cssCustomProperty});
+        margin-bottom: 0px;
+        "
+    >
+        appearance: ${() => appearanceLabel}
+    </p>
+    <${chatConversationTag} appearance="${() => appearance}">
+        <${chatMessageTag} message-type="inbound">
+            <span>Hello.</span>
+        </${chatMessageTag}>
+        <${chatMessageTag} message-type="outbound">
+            <span>Greetings!</span>
+        </${chatMessageTag}>
+        <${chatInputTag} slot='input'></${chatInputTag}>
+    </${chatConversationTag}>
+`;
+
+export const conversationAppearance: StoryFn = createMatrixThemeStory(html`
+    ${createMatrix(conversationWithAppearance, [appearanceStates])}
 `);
 
 export const hidden: StoryFn = createStory(